Refactor: 회원가입/로그인 레이아웃 컴포넌트화
